Atencion Medica pdf formato 20260711 CN

+++ src/main/java/com/api_salud/api_salud/service/PdfGeneratorService.java
@@ -15,5 +15,5 @@ public interface PdfGeneratorService {
         String nombreMedico*/
     );
     
-    byte[] estamparRubricaMedico(byte[] pdfBytes, Integer idMedico, String nombreMedico, String cmpMedico);
+    byte[] estamparRubricaMedico(byte[] pdfBytes, Long idMedico, String nombreMedico, String cmpMedico);
 }
